Se convierten pagos a decimales, ajustando validaciones, controllers y logica en edicion y creacion de pagos mediante pruebas con postman. Se reduce la base de datos quitando imagenes que pasaban a ser innecesarias, y se cambia build a develop para realizar las ultimas pruebas sin afectar heroku en demo.

+++ db/models/Project.js
@@ -15,11 +15,11 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: true,
     },
     total: {
-      type: DataTypes.BIGINT.UNSIGNED,
+      type: DataTypes.DECIMAL.UNSIGNED,
       allowNull: false,
     },
     balance: {
-      type: DataTypes.BIGINT.UNSIGNED,
+      type: DataTypes.DECIMAL.UNSIGNED,
       allowNull: true,
     },
     cashflow: {
